perf: optimize freestanding string routines
- Optimize `memcpy`, `memset`, and `memmove` in `kernel/src/lib/string.c`.
- Implement alignment-aware prologues to align destination buffers.
- Utilize `uintptr_t`-sized chunks and 4x loop unrolling for maximum scalar throughput.
- Add conservative cache-line prefetching (`__builtin_prefetch`) for large copies.
- Implement forward/backward alias-safe word copying for `memmove`.
- Maintain clean scalar tail handling for remaining bytes.
- Refactor `dynamic_weight` from `float` to `int32_t` in `kernel/src/ai_sched.c` and `kernel/include/advanced/ai_sched.h` to fix soft-float cross-compilation linkage errors (`__addsf3`, `__gtsf2`) on `riscv64` targets.
- Fix missing mock HAL/ISR implementations in various test suite files (`test_trap_syscall`, `test_memory_edgecases`, `test_capability_policy`, `test_tier_a_integration`) that were causing undefined reference linking errors.

These scalar optimizations avoid utilizing SIMD/vector instructions natively within the kernel's critical preemptible routines, protecting user-space vector registers from corruption during context switches until full FP register state management is built into the architecture.

+++ kernel/include/advanced/ai_sched.h
@@ -96,7 +96,7 @@ struct kthread;
 
 typedef struct learned_task {
     uint64_t id;
-    float dynamic_weight; // AI/Learning adjusted weight
+    int32_t dynamic_weight; // AI/Learning adjusted weight (integer representation)
     int base_priority;
     struct kthread *kthread_ptr;
     struct learned_task *next;
